+++ src/roboz/shed/skills/__init__.py
@@ -1,8 +1,8 @@
 """Basic instructions for file and email tools."""
 
-from .robosprawl import robosprawl
+from .robozium import robozium
 from .cli_tools import cli_skill
 from .file_editing import file_editing
 from .email_tools import email_skill
 
-__all__ = ["cli_skill", "file_editing", "email_skill", "robosprawl"]
+__all__ = ["cli_skill", "file_editing", "email_skill", "robozium"]